Refactor the Html Datatype Class into text.py instead of images.py
