+++ lib/galaxy/datatypes/text.py
@@ -12,9 +12,49 @@ import tempfile
 
 from galaxy.datatypes.data import get_file_peek, Text
 from galaxy.datatypes.metadata import MetadataElement, MetadataParameter
+from galaxy.datatypes.sniff import get_headers
 from galaxy.util import nice_size, string_as_bool
 
 log = logging.getLogger(__name__)
+
+
+class Html( Text ):
+    """Class describing an html file"""
+    edam_format = "format_2331"
+    file_ext = "html"
+
+    def set_peek( self, dataset, is_multi_byte=False ):
+        if not dataset.dataset.purged:
+            dataset.peek = "HTML file"
+            dataset.blurb = nice_size( dataset.get_size() )
+        else:
+            dataset.peek = 'file does not exist'
+            dataset.blurb = 'file purged from disk'
+
+    def get_mime(self):
+        """Returns the mime type of the datatype"""
+        return 'text/html'
+
+    def sniff( self, filename ):
+        """
+        Determines whether the file is in html format
+
+        >>> from galaxy.datatypes.sniff import get_test_fname
+        >>> fname = get_test_fname( 'complete.bed' )
+        >>> Html().sniff( fname )
+        False
+        >>> fname = get_test_fname( 'file.html' )
+        >>> Html().sniff( fname )
+        True
+        """
+        headers = get_headers( filename, None )
+        try:
+            for i, hdr in enumerate(headers):
+                if hdr and hdr[0].lower().find( '<html>' ) >= 0:
+                    return True
+            return False
+        except:
+            return True
 
 
 class Json( Text ):
